Fix all the things
JavaScript is such a mess.  These changes fix jQuery errors introduced
by... 1.11.3?  1.12.0?  I don't know.  They happened in 1.11.0 before
I upgraded it, so maybe I don't need to replace it with 1.11.3 (but
this works now, so I'm not doing any more experimenting).  But
apparently they were never noticed because of the Org HTML export IDs
being hex strings and never having "%20" in them, which my
unpackaged/org-export-html-with-useful-ids-mode does.  So this fixes
that problem, which was causing JavaScript errors and breaking the
mini-TOC.

But these changes, in particular, I don't even understand.  I tried to
make a single escape_jquery_chars function that would escape both "#"
and "%", but it did not fix the problems.  I only managed to fix them
by doing these specific escapes in these specific places in these
specific ways.  I cannot explain why.  I tried to do the sensible
thing, but of course JavaScript (or jQuery, who knows) said, "Nope."
Only with the right magic incantations does it work.  I got lucky with
trial-and-error, and this works, so I'm not messing with it anymore.